fix: Allow case-insensitive matching by introducing new StrEnum subclasses and trafarets
